+++ R/pathsToLines.R
@@ -26,9 +26,9 @@ pathsToLines <- function(sim){
   er <- sim$costSurface == 0 
   
   linelist <- lapply(1:length(sim$paths.list), function(i){
-    
+   
     # finds first match for start and end cells in paths
-    inds <- match(sim$paths.list[[i]], sim$paths.v$V1)
+    inds <- match(sim$paths.list[[i]], sim$paths.v[[1]])
     
     if(any(is.na(inds))){
       stop("NA values in cost raster along paths, check raster", call. = FALSE)
@@ -38,13 +38,16 @@ pathsToLines <- function(sim){
       return(NULL)
     }
     # cell indicies for vertices on line
-    v <- sim$paths.v$V1[inds[1]:inds[2]]
+    v <- sim$paths.v[[1]][inds[1]:inds[2]]
     
     # remove vertices in this line from sim object in parent environment
     sim$paths.v <<- sim$paths.v[-(inds[1]:inds[2]), ]
     
+    # raster values at v
+    er_v <- terra::extract(er, v)[[1]]
+    
     ## index of where new road connects to existing road
-    conn <- which(er[v] == 1)
+    conn <- which(er_v == 1)
     
     # outLine1 <- sf::st_linestring(raster::xyFromCell(sim$costSurface, v))
     
@@ -55,15 +58,15 @@ pathsToLines <- function(sim){
     ## something with cumsum(rle(er[v]==1)$lengths)?
     if(length(conn) > 0){
     
-      if(length(which(er[v] == 0)) == 0){
+      if(length(which(er_v == 0)) == 0){
         # the whole path is on existing road
         return(NULL)
       }
       
-      run_lengths <- rle(er[v]==0)$length 
+      run_lengths <- rle(er_v == 0)$length 
       if(length(run_lengths) > 1){
-        if(er[v[1]] == 0){
-          if(er[v[length(v)]] == 1){
+        if(er_v[1] == 0){
+          if(er_v[length(v)] == 1){
             # first cell is not existing road and last cell is existing road
             run_lengths_mat <- run_lengths %>% cumsum() 
             run_lengths_mat <- c(1, run_lengths_mat)
@@ -82,7 +85,7 @@ pathsToLines <- function(sim){
               mutate(end = .data$end + 1)
           }
         } else {
-          if(er[v[length(v)]] == 1) {
+          if(er_v[length(v)] == 1) {
             # first cell is existing road and last cell is existing road
             run_lengths_mat <- run_lengths %>% cumsum() 
             run_lengths_mat <- run_lengths_mat[-length(run_lengths_mat)] %>% 
@@ -103,7 +106,8 @@ pathsToLines <- function(sim){
           keep <- list(seq(from = run_lengths_mat$start, 
                        to = run_lengths_mat$end))
         } else {
-          keep <- apply(run_lengths_mat, 1, function(x) seq(from = x[1], to = x[2]))
+          keep <- apply(run_lengths_mat, 1, function(x) seq(from = x[1], to = x[2]), 
+                        simplify = FALSE)
         } 
       }
       
@@ -113,14 +117,14 @@ pathsToLines <- function(sim){
     if(length(keep) == 1){
       cellsToKeep <- v[keep[[1]]]
       cellsToKeep <- na.omit(cellsToKeep)
-      outLine <- sf::st_linestring(raster::xyFromCell(sim$costSurface, 
+      outLine <- sf::st_linestring(terra::xyFromCell(sim$costSurface, 
                                                       cellsToKeep)) %>%
         sf::st_sfc()
     } else {
       outLine <- lapply(keep, function(x){
         cellsToKeep <- v[x]
         cellsToKeep <- na.omit(cellsToKeep)
-        sf::st_linestring(raster::xyFromCell(sim$costSurface, cellsToKeep))
+        sf::st_linestring(terra::xyFromCell(sim$costSurface, cellsToKeep))
       })
       outLine <- sf::st_union(outLine %>% sf::st_as_sfc())
     }
@@ -144,6 +148,10 @@ pathsToLines <- function(sim){
   
   # remove NULLs
   linelist <- linelist[vapply(linelist,function(x) !is.null(x), c(TRUE))]
+  
+  if(length(linelist) == 0){
+    return(slice(sim$roads, 0))
+  }
   
   outLines <- do.call(c, linelist) %>% 
     sf::st_union()